requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
hf-xet==1.2.0
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.5.0
Jinja2==3.1.6
joblib==1.4.2
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai-whisper==20250625
packaging==26.0
pandas==2.0.3
pillow==10.4.0
propcache==0.2.0
protobuf==5.29.6
psutil==7.2.2
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.3
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.2.3
xxhash==3.6.0
yarl==1.15.2
zipp==3.20.2
